Removed `Require cycles are allowed` warning
